refactor: move internal upload manager to own module
